Separate text tags into their own module
