+++ GTG/gtk/editor/taskview2.py
@@ -16,18 +16,22 @@
 # this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------
 
+"""Taskview widget used in the task editor."""
+
+
 import re
 from time import time
 
-from gi.repository import Gtk, Pango, GLib, Gdk
+from gi.repository import Gtk, GLib, Gdk
 
 from GTG.core.logger import log
 from GTG.core.requester import Requester
-from GTG.gtk.colors import background_color
 import GTG.core.urlregex as url_regex
 from webbrowser import open as openurl
 from gettext import gettext as _
-from GTG.core.task import Task
+
+from GTG.gtk.editor.text_tags import (TitleTag, SubTaskTag, InvisibleTag,
+                                      TaskTagTag, InternalLinkTag, LinkTag)
 
 
 # Regex to find GTG's tags
@@ -38,250 +42,6 @@ TAG_REGEX = re.compile(r'\@\w+')
 INTERNAL_REGEX = re.compile((r'gtg:\/\/'
                              r'[0-9a-f]{8}(?:-[0-9a-f]{4}){3}-[0-9a-f]{12}'),
                             re.IGNORECASE)
-
-
-class SubTaskTag(Gtk.TextTag):
-    """Subtask Text tag."""
-
-
-    def __init__(self, task: Task) -> None:
-        super().__init__()
-
-        self.tid = task.tid
-
-        self.set_property('background', 'white')
-        self.set_property('underline', Pango.Underline.SINGLE)
-        self.set_property('left-margin', 40)
-
-        if task.status == Task.STA_ACTIVE:
-            self.set_property('strikethrough', False)
-            self.set_property('foreground', '#007bff')
-        else:
-            self.set_property('strikethrough', True)
-            self.set_property('foreground', 'gray')
-
-        self.connect('event', self.on_tag)
-
-
-    def on_tag(self, tag, view, event, _iter) -> None:
-        """Callback for events that happen inside the tag."""
-
-        button = event.get_button()
-
-        # If there was a click...
-        if button[0] and button[1] == 1:
-            view.open_subtask(self.tid)
-
-
-    def activate(self, view) -> None:
-        """Open the link in this tag."""
-
-        view.open_subtask(self.tid)
-
-
-    def set_hover(self) -> None:
-        """Change tag appareance when hovering."""
-
-        self.set_property('background', 'light gray')
-
-
-    def reset(self) -> None:
-        """Reset tag appareance when not hovering."""
-
-        self.set_property('background', 'white')
-
-
-class InvisibleTag(Gtk.TextTag):
-    """Subtask Text tag."""
-
-
-    def __init__(self) -> None:
-        super().__init__()
-
-        self.set_property('invisible', True)
-        self.set_property('left-margin', 40)
-        self.set_property('foreground', '#111111')
-        self.set_property('background', '#DDDDDD')
-
-
-    def set_hover(self) -> None:
-        """Change tag appareance when hovering."""
-
-        self.set_property('invisible', False)
-
-
-    def reset(self) -> None:
-        """Reset tag appareance when not hovering."""
-
-        self.set_property('invisible', True)
-
-
-class InternalLinkTag(Gtk.TextTag):
-    """Internal Link Text tag (for urls)."""
-
-
-    def __init__(self, task: Task) -> None:
-        super().__init__()
-
-        self.tid = task.tid
-
-        self.set_property('background', 'white')
-        self.set_property('underline', Pango.Underline.SINGLE)
-
-        if task.status == Task.STA_ACTIVE:
-            self.set_property('strikethrough', False)
-            self.set_property('foreground', '#007bff')
-        else:
-            self.set_property('strikethrough', True)
-            self.set_property('foreground', 'gray')
-
-        self.connect('event', self.on_tag)
-
-
-    def on_tag(self, tag, view, event, _iter) -> None:
-        """Callback for events that happen inside the tag."""
-
-        button = event.get_button()
-
-        # If there was a click...
-        if button[0] and button[1] == 1:
-            view.open_subtask(self.tid)
-
-
-    def activate(self, view) -> None:
-        """Open the link in this tag."""
-
-        view.open_subtask(self.tid)
-
-    def set_hover(self) -> None:
-        """Change tag appareance when hovering."""
-
-        self.set_property('background', 'light gray')
-
-
-    def reset(self) -> None:
-        """Reset tag appareance when not hovering."""
-
-        self.set_property('background', 'white')
-
-
-class LinkTag(Gtk.TextTag):
-    """Link Text tag (for urls)."""
-
-
-    def __init__(self, url: str) -> None:
-        super().__init__()
-
-        self.url = url
-
-        self.set_property('background', 'white')
-        self.set_property('foreground', '#007bff')
-        self.set_property('underline', Pango.Underline.SINGLE)
-        self.set_property('strikethrough', False)
-
-        self.connect('event', self.on_tag)
-
-
-    def on_tag(self, tag, view, event, _iter) -> None:
-        """Callback for events that happen inside the tag."""
-
-        button = event.get_button()
-
-        # If there was a click...
-        if button[0]:
-
-            # Left click
-            if button[1] == 1:
-                openurl(self.url)
-
-            # Right click
-            elif button[1] == 3:
-                view.clicked_link = self.url
-
-
-    def activate(self, view) -> None:
-        """Open the link in this tag."""
-
-        openurl(self.url)
-
-
-    def set_hover(self) -> None:
-        """Change tag appareance when hovering."""
-
-        self.set_property('background', 'light gray')
-
-
-    def reset(self) -> None:
-        """Reset tag appareance when not hovering."""
-
-        self.set_property('background', 'white')
-
-
-class TitleTag(Gtk.TextTag):
-    """Title Text tag (only one per buffer)."""
-
-
-    def __init__(self) -> None:
-        super().__init__()
-
-        self.set_property('weight', Pango.Weight.BOLD)
-        self.set_property('size_points', 16)
-        self.set_property('pixels_above_lines', 15)
-        self.set_property('pixels_below_lines', 30)
-
-
-class TaskTagTag(Gtk.TextTag):
-    """Text tag for task tags."""
-
-
-    def __init__(self, tag: str, req: Requester) -> None:
-        super().__init__()
-
-        self.tag_name = tag
-        self.tag = req.get_tag(tag)
-
-        try:
-            self.color = background_color([self.tag]) or'#FFEA00'
-        except AttributeError:
-            self.color = '#FFEA00'
-
-        self.set_property('background', self.color)
-        self.set_property('foreground', 'black')
-
-        self.connect('event', self.on_tag)
-
-
-    def set_hover(self) -> None:
-        """Change tag appareance when hovering."""
-
-        try:
-            color = self.tag.get_attribute('color') or '#EBDB34'
-            self.set_property('background', color)
-        except AttributeError:
-            self.set_property('background', '#EBDB34')
-
-
-    def reset(self) -> None:
-        """Reset tag appareance when not hovering."""
-
-        self.set_property('background', self.color)
-
-
-    def on_tag(self, tag, view, event, _iter) -> None:
-        """Callback for events that happen inside the tag."""
-
-        button = event.get_button()
-
-        if button[0] and button[1] == 1:
-            view.browse_tag(self.tag_name)
-
-
-    def activate(self, view) -> None:
-        """Open the link in this tag."""
-
-        view.browse_tag(self.tag_name)
-
-
 
 
 class TaskView(Gtk.TextView):
